% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_labels.R
\name{sl_label_page}
\alias{sl_label_page}
\title{Create a tibble which defines the dimensions of sticky labels on a page}
\usage{
sl_label_page(paper_size, label_size, margin, padding)
}
\value{
a 'stickylabels' tibble
}
\description{
Returns a tibble in which each observation corresponds to one sticky label, and variables define dimensions
}
