% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_labels.R
\name{sl_labels}
\alias{sl_labels}
\title{Create a tibble which defines the dimensions of sticky labels on a page}
\usage{
sl_labels(paper_size, label_size, margin, padding, units = "mm")
}
\arguments{
\item{paper_size}{A string passed to sl_paper() (eg. "A4", "letter", etc.) or a length-two numeric vector specifying paper height and width in \code{units}.}

\item{label_size}{A length-two numeric vector specifying label height and width in \code{units}.}

\item{margin}{A length-two numeric vector specifying the horizontal and vertical page margin (unused space at the edge of the page).}

\item{padding}{A length-two numeric vector specifying the horizontal and vertical padding between each label.}

\item{units}{A character string passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}. One of \verb{c("mm", "cm" "in", "px")}.}
}
\value{
a 'stickylabels' tibble
}
\description{
Returns a tibble in which each observation corresponds to one sticky label, and variables define dimensions
}
