% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_page.R
\name{sl_page}
\alias{sl_page}
\title{Specify a paper size for a page of labels}
\usage{
sl_page(page_specification)
}
\value{
page_specification An UK paper size from "A1" to "A5", or any of "tabloid", "legal", "letter" or "half letter"
}
\description{
This function takes a character string, such as 'A4' or 'letter' and returns the dimensions of the paper size in inches. To be used as a helper in functions like sl_labels()
}
